refactor : 나머지 member service 이동
issue : #191
