+++ src/main/java/keeper/project/homepage/user/service/posting/CommentService.java
@@ -16,8 +16,8 @@ import keeper.project.homepage.repository.member.MemberHasCommentLikeRepository;
 import keeper.project.homepage.common.repository.posting.CommentRepository;
 import keeper.project.homepage.common.repository.posting.CommentSpec;
 import keeper.project.homepage.common.repository.posting.PostingRepository;
-import keeper.project.homepage.service.member.MemberHasCommentDislikeService;
-import keeper.project.homepage.service.member.MemberHasCommentLikeService;
+import keeper.project.homepage.user.service.member.MemberHasCommentDislikeService;
+import keeper.project.homepage.user.service.member.MemberHasCommentLikeService;
 import keeper.project.homepage.service.posting.PostingService;
 import keeper.project.homepage.user.service.member.MemberService;
 import lombok.RequiredArgsConstructor;
