+++ src/test/java/keeper/project/homepage/ApiControllerTestSetUp.java
@@ -28,8 +28,8 @@ import keeper.project.homepage.repository.ThumbnailRepository;
 import keeper.project.homepage.repository.member.MemberRepository;
 import keeper.project.homepage.admin.service.library.BookManageService;
 import keeper.project.homepage.user.service.attendance.AttendanceService;
-import keeper.project.homepage.service.member.MemberHasCommentDislikeService;
-import keeper.project.homepage.service.member.MemberHasCommentLikeService;
+import keeper.project.homepage.user.service.member.MemberHasCommentDislikeService;
+import keeper.project.homepage.user.service.member.MemberHasCommentLikeService;
 import keeper.project.homepage.user.service.posting.CommentService;
 import keeper.project.homepage.common.service.sign.SignUpService;
 import keeper.project.homepage.user.service.member.MemberService;
